# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: nebula-mgmt API
  version: "1.0.0"
  description: >
    Control-plane REST API consumed by the CLI, the nebula-agent poller, and
    mobile bundles. Agent endpoints (enroll, agent/updates) are unauthenticated
    at the bearer layer — enroll carries a single-use token, polls carry an
    Ed25519 proof-of-possession signature (ADR 0004). Admin endpoints require an
    operator API key.
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /api/v1/hosts/{id}:
    get:
      operationId: getHost
      summary: Get a host by id.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Host.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Host"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Operator API key as a bearer token.
  responses:
    Error:
      description: Error envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Host:
      type: object
      required:
        - id
        - network_id
        - name
        - nebula_ips
        - groups
        - role
        - is_lighthouse
        - is_relay
        - status
        - kind
        - created_at
        - updated_at
      properties:
        id:
          type: string
        network_id:
          type: string
        ca_id:
          type: string
        name:
          type: string
        nebula_ips:
          type: array
          items:
            type: string
        groups:
          type: array
          items:
            type: string
        role:
          type: string
        is_lighthouse:
          type: boolean
        is_relay:
          type: boolean
        public_ip:
          type: string
        listen_port:
          type: integer
        status:
          type: string
        cert_fingerprint:
          type: string
        prev_cert_fingerprint:
          type: string
        cert_expires_at:
          type: string
          format: date-time
        cert_rotated_at:
          type: string
          format: date-time
        pending_rekey:
          type: boolean
        signing_pub_pem:
          type: string
        last_seen_at:
          type: string
          format: date-time
        kind:
          type: string
        variant:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...